Make the `lock_api` dependency optional.
Without `lock_api`, this crate only provides a small number of types
such as `RawMutex` and `RawRwLock`, but if someone needs just those,
it's reasonable to provide them without depending on `lock_api`.

This also makes `RawMutex` and `RawRwLock` easier to use, because
it's no longer necessary to import the `lock_api` traits with the `as _`
trick.